fix property-scoped search
All property-scoped searches failed with `Error parsing query Syntax Error`.
This happened because the following code

    let full_query = if let Some(prop) = &params.property {
        format!("property:{:?} AND {}", prop, &q)
    } else {
        q
    };

would produce a query like:
https://atomicdata.dev/properties/name:hello

and this likely fails to parse because of multiple ":" in a string.

You can test this with the following link:
https://atomicdata.dev/search?q=hello&property=https%3A%2F%2Fatomicdata.dev%2Fproperties%2Fname

I am not sure how this was supposed to work, but I intuited my way to
add property:"property-name" to query and AND-combine that with user's
query.

    format!("property:{:?} AND {}", prop, &q)

I could probably scope user's query to value:

    format!("property:{:?} AND value:{}", prop, &q)

but I decided not to do this, so that the user could query by subject,
too.