Add version to package.json
This should help with some of the tools we are trying to use to package
this for Summer of Nix as part of https://github.com/ngi-nix/ngipkgs/issues/15

One unclear thing is that https://github.com/atomicdata-dev/atomic-server/releases/tag/v0.34.3
is tagged in git as 0.34.3 but the title of the release is 0.34.5. Not
sure if that is a typo, but we went with the git tag.

Co-authored-by: Jason Odoom <jasonodoom@gmail.com>
Co-authored-by: Anish Lakhwara <anish+git@lakhwara.com>
Co-authored-by: Dominic Mills <dominic.millz27@gmail.com>
Co-authored-by: Albert Chae <albertchae@users.noreply.github.com>
Co-authored-by: Jack Leightcap <jack@leightcap.com>