Add flush-to-disk destructor for Db
Without this, any data we have added is only going to be properly
persisted if sled happened to complete one of its automatic syncs after
the addition. When running a server, this is usually not an issue, but
a command like `atomic-server import` may complete quite quickly, making
us lose some or even all of our data to be imported.